MVVM model
Přidáno: MVVM model zpracování, obrana při boji, použití předmětu, vylepšen nákup v marketu

+++ Draci_doupe/Draci_doupe/Interface/IProgressBarValue.cs
@@ -10,6 +10,6 @@ namespace Draci_doupe.Interface
     {
         int Maximum { get; set; }
         int Minimum { get; set; }
-        int Progress { get; set; }
+        double Progress { get; set; }
     }
 }
